+++ Coliseum/Assets/Scripts/Will/RangedEnemy.cs
@@ -64,7 +64,7 @@ public class RangedEnemy : MonoBehaviour
         // The above is to be removed
     }
 
-    public void Knockback(int kb) {
+    public void Knockback(float kb) {
         enemyRb.AddForce(-1 * direction * new Vector2(kb, kb));
         timer += 1f;
     }
